사용자의 Dailyinfo 출력시 날짜가 중복되어 출력되는 버그 픽스

+++ src/univHealth/UserFood.java
@@ -8,7 +8,7 @@ public class UserFood {
 	public UserFood(Food food, int size) {
 		this.food = food;
 		this.size = size;
-		this.cal = food.getCalRatio() * size;
+		this.cal = (int) (food.getCalRatio() * size);
 	}
 
 	@Override
